biglake-iceberg-pipeline/services/data-cleaning-agent/pyproject.toml
[project]
name = "datagrunt-agent"
version = "0.1.0"
description = "A data engineering agent for universal file ingestion, schema detection, schema evolution, and canonical schema transformation."
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.5.0",
    "duckdb>=1.4.0",
    "polars>=1.7.1",
    "python-dotenv",
    "functions-framework>=3.0.0",
    "google-cloud-storage>=2.0.0",
    "google-cloud-pubsub>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
